The hard margin is got by the ppd_size_t information
git-svn-id: https://svn.code.sf.net/p/splix/code/splix@211 10358e0b-e31b-0410-b13d-bf5c54612e97

+++ include/printer.h
@@ -51,8 +51,8 @@ class Printer
         float                   _pageWidth;
         float                   _pageHeight;
 
-        unsigned long           _hardMarginX;
-        unsigned long           _hardMarginY;
+        float                   _hardMarginX;
+        float                   _hardMarginY;
 
     public:
         /**
@@ -144,11 +144,11 @@ class Printer
         /**
          * @return the hard margin of the printer in the X direction.
          */
-        unsigned long           hardMarginX() const {return _hardMarginX;}
+        float                   hardMarginX() const {return _hardMarginX;}
         /**
          * @return the hard margin of the printer in the Y direction.
          */
-        unsigned long           hardMarginY() const {return _hardMarginY;}
+        float                   hardMarginY() const {return _hardMarginY;}
 
 };
